# [nvim] Migrate from Packer to Lazy