--- a/config/nvim/lua/keys.lua
+++ b/config/nvim/lua/keys.lua
@@ -21,8 +21,6 @@ utils.map('n', '<leader>ps', [[<cmd>PackerSync<cr>]], { silent = true }) -- Run 
 --" of the line on second press. It alternates afterwards.
 utils.map('n', '0', "virtcol('.') - 1 <= indent('.') && col('.') > 1 ? '0' : '_'", { expr = true })
 
-utils.map('n', '<leader>e', '<cmd>NvimTreeToggle<cr>', { silent = true })
-
 -- Telescope Bindings
 utils.map('n', '<C-p>', '<cmd>Telescope find_files<cr>')
 utils.map('n', '<leader>;', '<cmd>Telescope buffers<cr>')