--- a/config/nvim/lua/plugins/init.lua
+++ b/config/nvim/lua/plugins/init.lua
@@ -95,20 +95,6 @@ return {
   },
 
   {
-    'nvim-lualine/lualine.nvim',
-    dependencies = {
-      'kyazdani42/nvim-web-devicons', opt = true
-    },
-    config = function()
-      require('lualine').setup {
-        options = {
-          theme = 'palenight',
-        },
-      }
-    end
-  },
-
-  {
     'lewis6991/gitsigns.nvim', -- Adds Git highlighting
     dependencies = {
       'nvim-lua/plenary.nvim'