--- a/config/nvim/lua/plugins/lualine.lua
+++ b/config/nvim/lua/plugins/lualine.lua
@@ -1,0 +1,11 @@
+return {
+  'nvim-lualine/lualine.nvim',
+  dependencies = {
+    'kyazdani42/nvim-web-devicons', opt = true
+  },
+  config = {
+    options = {
+      theme = 'palenight',
+    },
+  },
+}