Migrates from Packer to Lazy for a better plugin management experience in NeoVim

## Performance Delta
I had previously set up Packer to lazy load a handful of modules. Otherwise the performance would be much starker

### Before
```
Benchmark 1: nvim --headless +qa
  Time (mean ± σ):     117.8 ms ±   1.0 ms    [User: 73.1 ms, System: 40.1 ms]
  Range (min … max):   116.1 ms … 119.9 ms    24 runs
```

### After 

```
Benchmark 1: nvim --headless +qa
  Time (mean ± σ):      91.2 ms ±   1.7 ms    [User: 57.0 ms, System: 32.0 ms]
  Range (min … max):    86.3 ms …  95.1 ms    31 runs
```

Roughly a 30% gain